+++ src/main/java/ch/hsr/graph/model/Edge.java
@@ -1,7 +1,7 @@
 package ch.hsr.graph.model;
 
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
-import ch.hsr.adv.lib.graph.logic.domain.ADVEdge;
+import ch.hsr.adv.commons.graph.logic.domain.ADVEdge;
 
 public class Edge<E> implements ADVEdge<E> {
