+++ src/main/java/ch/hsr/graph/model/Vertex.java
@@ -1,8 +1,8 @@
 package ch.hsr.graph.model;
 
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
-import ch.hsr.adv.lib.graph.logic.domain.ADVGraph;
-import ch.hsr.adv.lib.graph.logic.domain.ADVVertex;
+import ch.hsr.adv.commons.graph.logic.domain.ADVGraph;
+import ch.hsr.adv.commons.graph.logic.domain.ADVVertex;
 
 import java.util.concurrent.atomic.AtomicInteger;
